backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
greenlet==3.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0

# Async Tasks
celery==5.3.4
redis==5.0.1

# File Storage
boto3==1.34.0
minio==7.2.0

# Search
elasticsearch==8.11.0

# Scientific Computing (for ML metrics)
numpy==1.24.4
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.4

# Explainability
shap==0.43.0
lime==0.2.0.1

# Fairness Metrics
fairlearn==0.10.0
aif360==0.5.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
tenacity==8.2.3
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1